requirements.txt
Flask==2.0.1
opencv-python==4.5.3.20210927
numpy==1.21.2
tensorflow==2.6.0
mysql-connector-python==8.0.26